pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
]
